requirements.txt
openai~=1.59.3
python-dotenv~=1.0.1
resend~=2.5.1
pydantic[email]~=2.10.4